# Speed up Rspec tests