--- a/.env.test
+++ b/.env.test
@@ -3,3 +3,4 @@ NODE_ENV=tests
 # Federation
 LOCAL_DOMAIN=cb6e6126.ngrok.io
 LOCAL_HTTPS=true
+GITHUB_REPOSITORY=TheBLVD/moth.social