--- a/.gitignore
+++ b/.gitignore
@@ -61,3 +61,4 @@ yarn-debug.log
 
 # Ignore Docker option files
 docker-compose.override.yml
+/backups/