--- a/config/application.rb
+++ b/config/application.rb
@@ -45,7 +45,11 @@ require_relative '../lib/active_record/database_tasks_extensions'
 require_relative '../lib/active_record/batches'
 require_relative '../lib/simple_navigation/item_extensions'
 
-Dotenv::Railtie.load
+if 'test' == ENV['RAILS_ENV']
+  Dotenv::Railtie.overload
+else
+  Dotenv::Railtie.load
+end
 
 Bundler.require(:pam_authentication) if ENV['PAM_ENABLED'] == 'true'