The connection pool reaper checks for idle connections every 30 seconds, and an idle connection is one that's been idle for 30 seconds.  To test that, the original tests slept for 61 seconds, and then watched to see if the reaper ran.

I changed it so that for testing purposes, the reaper checks for idle connections every second, and then also moved the clock time to 100 seconds in the future.  Then I sleep for just two seconds, to give the reaper a chance to run, and see if it did.